Consolidate duplicate tests and clean up dependencies
- Merged test_graph_token.py into test_graph.py to eliminate ~80% test duplication
- Added connection_name parameter testing to consolidated tests
- Removed unused dotenv dependency from tests/graph/pyproject.toml
- All 17 tests pass with comprehensive coverage maintained
