+++ tests/graph/pyproject.toml
@@ -5,9 +5,8 @@ description = "demo app to test Azure Graph integration"
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
-    "dotenv>=0.9.9",
-    "microsoft-teams-app",
-    "microsoft-teams-graph",
+    "microsoft-teams-app>=0.1.0",
+    "microsoft-teams-graph>=0.1.0",
 ]
 
 [tool.uv.sources]
